Problem with item drop log.
For some reason, a loot check is done multiple times in the DestroyEnemy coroutine. I need it to check one time. I've noted that this only started happening when I moved the code into the coroutine. I may take it out and use a bool to control the loot check in the update loop.