Loot object now displayed when item generated.
Initially, it wasn't working because I was trying to use the prefab instead of a separate game object, but when I fixed that problem it still wasn't showing up on screen. Not sure if it had to do with the sprite renderer being disabled on start (this was intentional) that somehow caused the entire object to not be active?